feat(page-cards-slider): base template
